Refactored code: Change folder name to camelCase, make font size sass varaibles
