--- src/components/Contacts/contacts.scss
@@ -1,32 +1,0 @@
-@import '../../style/base';
-
-.Contacts{
-
-  margin-top: 40px;
-
-  &>div{
-    margin-top: 10px;
-  }
-
-  &>div:first-child{
-    margin-top: 0px;
-  }
-  span{
-    display: inline-block;
-  }
-  .icons{
-    width: 12px;
-    color: $gray;
-  }
-
-  @include e(text-content){
-    margin-left: 10px;
-    flex: 3;
-    color: $orange;
-    a{
-      &:hover {
-        color: darken($color: $orange, $amount: 20%) !important;
-      }
-    }
-  }  
-}
