--- src/components/Profile/profile.scss
@@ -1,47 +1,0 @@
-@import '../../style/base';
-
-.Profile{
-
-  width: 100%;
-  overflow: hidden;
-  
-  @include mq(md){
-    width: 50%;
-    flex: 1;
-    max-width: 280px;
-  }
-
-  @include e (heading){
-    white-space: nowrap;
-    color: $ash-gray;
-  }
-
-  @include e(avatar){
-    border-radius: 50%;
-    width: 100%;
-  }
-
-  @include e(avatar-container){
-    margin: 0 auto;
-    margin-top: 32px;
-    max-width: 300px;
-  }
-
-  @include e(user-details){
-    margin-top: 40px;
-  }
-
-  @include e(user-id){
-    margin-top: 8px;
-  }
-
-  @include e(user-name){
-    color: $pear;
-    
-    a{
-      &:hover {
-        color: darken($color: $pear, $amount: 20%) !important;
-      }
-    }
-  }
-}
