--- src/components/Repositoires/RespositoryListItem/RepositoryListItem.scss
@@ -1,43 +1,0 @@
-@import '../../../style/base';
-
-.Repository-list{
-  margin-top: 64px;
-
-  &:first-child{
-    margin-top: 0px;
-  }
-
-  @include e(title){
-    color: $rose;
-    font-size: 18px;
-
-    a:hover{
-      color: $black;
-      background-color: $pacific-blue;
-    }
-  }
-
-  @include e(details){
-    margin-top: 13px;
-    display: flex;
-    &>div{
-      margin-right: 20px;
-    }
-  }
-
-  @include e(language){
-    color: $spray;
-    text-transform: capitalize;
-  }
-
-  @include e(description){
-    margin-top: 10px;
-    font-size: 13px;
-    color: $gray;
-  }
-
-  @include e(update-time){
-    color: $gray;
-  }
-
-}
