--- src/components/Socials/Socials.scss
@@ -1,25 +1,0 @@
-@import '../../style/base';
-
-.Socials{
-
-  margin-top: 20px;
-  display: flex;
-  @include e(followers){
-    flex:1;
-  }
-
-  @include e(following){
-    flex:1;
-  }
-
-  .spray{
-    color: $spray;
-    font-size: 16px;
-  }
-
-  a{
-    &:hover{
-      color: lighten($color: $spray, $amount: 20%);
-    }
-  }
-}
