+++ src/components/followers/followerListItem/followerListItem.scss
@@ -1,0 +1,38 @@
+@import '../../../style/base';
+
+.FollowerListItem{
+
+  margin-top: 48px;
+  display: flex;
+
+  &:first-child{
+    margin-top: 0px;
+  }
+
+  @include e(avatar-holder){
+    width: 60px;
+  }
+
+  @include e(avatar){
+    width:60px;
+    border-radius: 50%;
+  }
+
+  @include e(user-details){
+    margin-left: 28px;
+    position: relative;
+  }
+
+  @include e(user-id){
+    color: $rose;
+    font-size: $base__font-size;
+    position: absolute;
+    top: 20%;
+
+    a:hover{
+      color: $black;
+      background-color: $pacific-blue;
+    }
+  }
+
+}
