+++ src/components/repositoires/repositories.scss
@@ -1,0 +1,26 @@
+@import '../../style/base';
+
+.Repositories{
+  
+  overflow: hidden;
+  margin-top: 40px;
+
+  @include e (heading){
+    white-space: nowrap;
+    color: $ash-gray;
+  }
+
+  @include e(list-container){
+    margin-top: 32px;
+  }
+
+  @include mq(md){
+    flex: 2;
+    margin-left: 60px;
+    margin-top: 0px;
+  }
+
+  @include e(loading){
+    margin-top: 32px;
+  }
+}
