+++ src/components/repositoires/respositoryListItem/repositoryListItem.scss
@@ -1,0 +1,43 @@
+@import '../../../style/base';
+
+.Repository-list{
+  margin-top: 64px;
+
+  &:first-child{
+    margin-top: 0px;
+  }
+
+  @include e(title){
+    color: $rose;
+    font-size: $xl-font-size;
+
+    a:hover{
+      color: $black;
+      background-color: $pacific-blue;
+    }
+  }
+
+  @include e(details){
+    margin-top: 13px;
+    display: flex;
+    &>div{
+      margin-right: 20px;
+    }
+  }
+
+  @include e(language){
+    color: $spray;
+    text-transform: capitalize;
+  }
+
+  @include e(description){
+    margin-top: 10px;
+    font-size: $small-font-size;
+    color: $gray;
+  }
+
+  @include e(update-time){
+    color: $gray;
+  }
+
+}
