+++ src/style/partials/_config.scss
@@ -97,6 +97,7 @@ $palettes: (
 $default-font-size: $base__font-size - 2px;
 $small-font-size: $default-font-size - 2px;
 $lg-font-size: $default-font-size + 2px;
+$xl-font-size: $default-font-size + 4px;
 
 //Heading font size vairables
 $heading-line-height: 28px !default;
