+++ src/style/style.scss
@@ -2,7 +2,7 @@
 
 html,body{
   font-family: 'Conso', 'Courier New', monospace;
-  font-size: 14px;
+  font-size: $default-font-size;
   background-color: $base-color;
   color: $white;
 }
